checkpoint moved getTaggedWordsHTML to NounInverter
still broken, though
