+++ src/dataClasses/NounInverter.jsx
@@ -1,6 +1,18 @@
-// for a specific poem (lines of text)
+import * as R from "ramda";
+
+const spannedWord = (mainClass, extraClasses, lineNum, wordNum, word) =>
+  `<span class="${mainClass} ${extraClasses}" id="word_${lineNum}_${wordNum}">${word}</span>`;
+
+const punct = /([.,\/#!$%\^&\*;:{}=\-_`~()]+)/gm;
+const spacePunct = /([\s.,\/#!$%\^&\*;:{}=\-_`~()]+)/gm;
+const UNICODE_NBSP = "\u00A0";
+
+// for a specific parser and poem (lines of text)
 export class NounInverter {
-  constructor(poemTextLines) {
+  constructor(parser, poemTextLines) {
+    this.parser = parser;
+    this.poemTextLines = poemTextLines;
+
     this.falsePositiveCount = 0;
     this.falseNegativeCount = 0;
     // nounInverters are represented by jagged arrays of arrays indexed by
@@ -12,7 +24,7 @@ export class NounInverter {
     //    i.e. inverted means now should be considered not-a-noun if originally
     //      a noun or vice-versa
     this.rep = poemTextLines ? new Array(poemTextLines.length).fill([]) : [[]];
-    this._hasInitializedEveryLine = false;
+    this.recomputeNounOutlinesHTML(); // completes initialization
     console.log("NounInverter constructed");
   }
   // line, word are 1-based
@@ -41,13 +53,88 @@ export class NounInverter {
       // if (lineNum == 6) debugger;
     }
   }
-  get clonable() {
-    return this._hasInitializedEveryLine;
+  // Return the HTML of all the words tagged (as either noun or non-noun)
+  _getTaggedWordsHTML() {
+    // debugger;
+    const lines = this.poemTextLines;
+    const parser = this.parser;
+
+    if (!lines) {
+      return "";
+    }
+    let outlined = [];
+
+    const addNounSpans = (tagged, lineNum) => {
+      let mainClass;
+      let extraClasses;
+      let wordNum = 1; // 1-based
+
+      return tagged.map(([word, tag]) => {
+        extraClasses = "";
+        let nounTest = tag === "NN" || tag === "NNS";
+        // if (lineNum == 4 && wordNum == 2) debugger;
+        if (this.isInverted(lineNum, wordNum)) {
+          console.log(
+            `addNounSpans inverting lineNum ${lineNum} wordNum ${wordNum}`
+          );
+          nounTest = !nounTest;
+          extraClasses = "inverted";
+        }
+        mainClass = nounTest ? "noun" : "non-noun";
+        return spannedWord(mainClass, extraClasses, lineNum, wordNum++, word);
+      });
+    };
+
+    /* Algorithm
+      - match the spaces and punctuation, save that as matchedSpacePunct
+      - remove all the punctuation from the words and tag, saving in as
+        parser.tagWordsInLine
+      - Use the word count of each line to initialize each line of the NounInverter
+      - Even out the saved punctuation and lines for recombination
+      - Recombine the spaces and words in the right order, save as outlined
+      - mark the current NounInverter as cloneable now that initialization is complete
+    */
+    lines.forEach((line, index) => {
+      let lineNum = index + 1;
+
+      if (line === "") {
+        outlined.push("");
+        return;
+      }
+      const matchedSpacePunct = line
+        .match(spacePunct)
+        .map((s) => s.replace(/ /g, UNICODE_NBSP));
+
+      let taggedWords = parser.tagWordsInLine(line.replaceAll(punct, ""));
+      this.initLineIfNeeded(lineNum, taggedWords.length);
+
+      // even out the lengths of the two arrays for zipping
+      if (matchedSpacePunct.length < taggedWords.length) {
+        matchedSpacePunct.push("");
+      } else if (taggedWords.length < matchedSpacePunct.length) {
+        taggedWords.push(["", ""]);
+      }
+
+      // zip results in the correct order (i.e. starting with a space or not)
+      let first, second;
+      if (line[0].match(/\s/)) {
+        first = matchedSpacePunct;
+        second = addNounSpans(taggedWords, lineNum);
+      } else {
+        first = addNounSpans(taggedWords, lineNum);
+        second = matchedSpacePunct;
+      }
+      const recombined = R.unnest(R.zip(first, second)).join("");
+      outlined.push(recombined);
+    });
+    return outlined.join("<br>");
   }
-  // this should only set to true by AuthorData after every line of this
-  // NounInverter has been initialized.
-  set clonable(hasBeenInitialized) {
-    this._hasInitializedEveryLine = hasBeenInitialized;
+  recomputeNounOutlinesHTML() {
+    const textOuput = document.getElementById("text-output");
+    if (!textOuput) return;
+    textOuput.innerHTML = this.taggedWordsHTML = this._getTaggedWordsHTML();
+    // debugger;
+    return this.taggedWordsHTML;
   }
 }
